sipXivr: Fix null pointer exception in Voicemail forwarding
Oops.  No comment audio file caused an exception while trying to
determine the duration of said non exsistant file.

Fixes:
   XX-6453 - Unable to forward a inbox message without adding a comment.

--Woof!


git-svn-id: http://sipxecs.sipfoundry.org/rep/sipXecs/main@16360 ab1d8caa-1f67-47f1-9e81-24633a41865c

+++ sipXivr/src/main/java/org/sipfoundry/voicemail/Message.java
@@ -84,7 +84,7 @@ public class Message {
     }
     
     public void deleteTempWav() {
-        if (m_wavFile.exists()) {
+        if (m_wavFile != null && m_wavFile.exists()) {
             LOG.debug("Message:deleteTempWav deleting "+ m_wavFile.getPath());
             FileUtils.deleteQuietly(m_wavFile) ;
         }
@@ -144,14 +144,17 @@ public class Message {
 
     public long getDuration() {
         // Calculate the duration (in seconds) from the Wav file
-        try {
-            AudioInputStream ais = AudioSystem.getAudioInputStream(getWavFile());
-            float secs =  ais.getFrameLength() / ais.getFormat().getFrameRate();
-            m_duration = Math.round(secs); // Round up.
-        } catch (Exception e) {
-            String trouble = "Message::getDuration Problem determining duration of "+getWavFile().getPath();
-            LOG.error(trouble, e);
-            throw new RuntimeException(trouble, e);            
+        File wavFile = getWavFile();
+        if (wavFile != null) {
+            try {
+                AudioInputStream ais = AudioSystem.getAudioInputStream(wavFile);
+                float secs =  ais.getFrameLength() / ais.getFormat().getFrameRate();
+                m_duration = Math.round(secs); // Round up.
+            } catch (Exception e) {
+                String trouble = "Message::getDuration Problem determining duration of "+getWavFile().getPath();
+                LOG.error(trouble, e);
+                throw new RuntimeException(trouble, e);            
+            }
         }
         return m_duration;
     }
